CustomCursor migrated to React
