+++ index.html
@@ -77,7 +77,7 @@
     </script>
   </head>
   <body  class="laguna__home">
-    <div class="lagunacustom-cursor"></div>
+    
     <div id="laguna__animatedBG"></div>
 
     <div id="root"></div>
@@ -876,8 +876,6 @@
     </div>
     <div class="laguna__thumbnail__hoverImage"></div>
 
-    <!--<script src="src/js/laguna-color-themes.js"></script>-->
-    <script src="src/js/laguna-cursor.js"></script>
     <script src="src/js/laguna-overlays.js"></script>
     <script src="src/js/laguna-thumbnails.js"></script>
     <script src="src/js/laguna-scroll.js"></script>
