+++ src/App.jsx
@@ -1,14 +1,20 @@
-
+//Components
 import Header from './components/header'
-import ColorThemeToggle from './js/ColorThemeToogle'
+import CursorElement from './components/CursorElement'
 
+
+import ColorThemeToggle from './js/ColorThemeToogle'
+import CustomCursor from './js/CustomCursor'
 
 function App() {
 
   return (
     <> 
+      <CursorElement />
       <Header />
+
       <ColorThemeToggle />
+      <CustomCursor />
       {/* other components */}
 
     </>
