+++ src/components/CursorElement.jsx
@@ -1,0 +1,8 @@
+
+const CursorElement = () => {
+  return (
+    <div class="lagunacustom-cursor"></div>
+  );
+};
+
+export default CursorElement;
